refine _db_stored_procedure_code attributes name to index mapping
